refactor: move implementation of sde-names to compile package to avoid direct dependency on Model
