+++ packages/compile/src/index.js
@@ -1,7 +1,7 @@
 // Copyright (c) 2022 Climate Interactive / New Venture Fund
 
 export { generateCode } from './CodeGen.js'
-export { parseAndGenerate } from './Entry.js'
+export { parseAndGenerate, printNames } from './Entry.js'
 export { canonicalName } from './Helpers.js'
 export { createParser, parseModel } from './Parser.js'
 export { preprocessModel } from './Preprocessor.js'
